Lots of changes
* migrate to latest mongoDB driver 3.4.2
* customise data series names on charts
* IMPORTANT sensor value changes can now be persisted as often or as
rarely as necessary rather than the fixed 5 minute period
* new database schema that reduces storage space and increases
performance
* migration scripts to allow porting from old schema
* new sensor processing options on charts and queries - count,
boolean_count_rising_edge, boolean_count_falling_edge
* script periodicity now a separate field from triggers and with many
more possible values
* IMPORTANT on-demand sensors - new feature that populates a sensor on
demand by executing a script. On-demand sensors can be updated in value
at any point in time when their value is first requested and then it is
stored permanently in database. This is useful for values such as
maximum temperature in a day and other kinds of derivative values that
you may wish to chart or capture
* sensor info view to see storage requirements and other stats of
sensors in database. Button to delete data.